Renamed lint check tests to lowercase
