+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/KERNEL.cl
