+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/VHDL.cl
