+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/Verilog.cl
