+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/kernel.h
