+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/other_Verilog.cl
