+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/otherdir/vhdl.cl
