+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/otherthing.cl
