+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/some/dir/kernel.cl
