+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/some_kernel.cl
