+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/somedir/verilog.cl
