+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/thing.h
