+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/vERILOG.cl
