+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/vhdl.h
