+++ test/clang-tidy/Inputs/fpga-kernel-name-restriction/vhdl_number_two.cl
